+++ src/server/nautical/grammars/WindOrientedGrammar.cpp
@@ -3,7 +3,7 @@
  *      Author: Jonas Östlund <uppfinnarjonas@gmail.com>
  */
 
-#include <server/nautical/grammars/Grammar001.h>
+#include <server/nautical/grammars/WindOrientedGrammar.h>
 #include <server/common/string.h>
 #include <iostream>
 #include <server/math/hmm/StateAssign.h>
